cloud-brain/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "zuralog-cloud-brain"
version = "0.1.0"
description = "Zuralog Cloud Brain — AI Health Assistant Backend"

requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    "alembic>=1.14.0",
    "httpx>=0.28.0",
    "email-validator>=2.0.0",
    "slowapi>=0.1.9",
    "firebase-admin>=6.0.0",
    "openai>=1.60.0",
    "redis[hiredis]>=5.0.0",
    "celery[redis]>=5.4.0",
    "python-multipart>=0.0.9",
    "numpy>=2.0.0",
    "sentry-sdk[fastapi,celery,sqlalchemy,httpx]>=2.19.0",
    "upstash-redis>=1.4.0",
    "posthog>=3.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.1",
]
